+++ src/model.h
@@ -38,29 +38,24 @@ typedef float LegAngles[LEG_CONFIGURATION_SIZE];
 
 struct Darkpaw 
 {
-	vec3				foot_positons[LEGS];			// in body frame
-	bool				is_planted[LEGS];				// true if the leg is planted.
-	LegAngles			angles[LEGS];					// in radians
-	int					motor_positions[LEGS][LEG_MOTORS];
-	vec3				centre_of_pressure;		// in body frame
+    vec3       foot_positons[LEGS];			// in body frame
+    bool       is_planted[LEGS];			// true if the leg is planted.
+    LegAngles  angles[LEGS];			        // in radians
+    unsigned   motor_positions[LEGS][LEG_MOTORS];
+    vec3       centre_of_pressure;		        // in body frame
 };
 
 struct Darkpaw* new_model(void);
 
 void get_leg_position(enum Leg leg, LegAngles *angles, vec3 *out_array);
 
-void get_next_motor_position(struct Darkpaw *darkpaw, 
-							 vec2 body_velocity, 
-							 float body_angular_velocity,
-							 float height,
-						     float delta_time,
-						     short* motor_out[LEGS * LEG_MOTORS]);
-
-
-
-
-
-
+void get_next_motor_position(
+		struct Darkpaw *darkpaw, 
+		vec2 body_velocity, 
+		float body_angular_velocity,
+		float height,
+		float delta_time,
+		unsigned* motor_out[LEGS * LEG_MOTORS]);
 
 
 #endif // !MODEL_H
